requirements.txt
pyyaml
python-dotenv
duckdb
pandas
pyarrow
fsspec
rich

# Optional für Tests
cerberus
voluptuous
pytest
coverage
loguru
